+++ src/org/opensourcebim/services/IfcToMpgCollectionService.java
@@ -1,0 +1,33 @@
+package org.opensourcebim.services;
+
+import org.bimserver.bimbots.BimBotContext;
+import org.bimserver.bimbots.BimBotsException;
+import org.bimserver.bimbots.BimBotsInput;
+import org.bimserver.bimbots.BimBotsOutput;
+import org.bimserver.emf.IfcModelInterface;
+import org.bimserver.plugins.PluginConfiguration;
+import org.opensourcebim.ifccollection.MpgIfcObjectCollector;
+import org.opensourcebim.ifccollection.MpgObjectStore;
+
+public class IfcToMpgCollectionService extends IfcObjectCollectionBaseService {
+
+	@Override
+	public BimBotsOutput runBimBot(BimBotsInput input, BimBotContext bimBotContext, PluginConfiguration pluginConfiguration)
+			throws BimBotsException {
+
+		IfcModelInterface ifcModel = input.getIfcModel();
+
+		// Get properties from ifcModel
+		MpgIfcObjectCollector matParser = new MpgIfcObjectCollector();
+		MpgObjectStore ifcResults = matParser.collectIfcModelObjects(ifcModel);
+		
+		ifcResults.SummaryReport();
+		
+		return this.toBimBotsJsonOutput(ifcResults, "results object collection");
+	}
+
+	@Override
+	public String getOutputSchema() {
+		return "MPG_OBJECT_JSON_0_0_2";
+	}
+}
